Bug fixes in word detector

+++ cogs/word_detector.py
@@ -9,18 +9,15 @@ class WordDetectorCog(commands.Cog):
 		self.filename = '.wordfilter'
 	
 	@commands.Cog.listener()
-	async def on_message(self, message): 
-		if message.author.bot or str(message.author.top_role) == 'admins':
-			return
+	async def on_message(self, message):
 		async with self.lock:
-			words = open(self.filename, 'r')
-			for w in words.readlines():
-				if w.strip() in message.content.lower():
+			with open(self.filename, 'r') as readfile:
+				words = [line.strip() for line in readfile.readlines()]
+			for w in words:
+				if w.lower() in message.content.lower():
 					await message.delete()
-					if w == ':b:' or w == ':regional_indicator_b:':
-						await message.author.send('lol', delete_after=10.0)
-					else:
-						await message.author.send('Illegal word, message deleted', delete_after=10.0)
+					if not message.author.bot:
+						await message.author.send('Illegal word, message deleted', delete_after=30.0)
 					return
 	
 	@commands.command(
@@ -29,17 +26,23 @@ class WordDetectorCog(commands.Cog):
 		usage='addfilter <word1> [<word2>] ...'
 	)
 	async def addfilter(self, ctx, *words):
-		await ctx.message.delete()
+		found_duplicate = False
 		if str(ctx.author.top_role) != 'admins':
-			await ctx.author.send('This command is for admins only', delete_after=10.0)
+			await ctx.author.send('This command is for admins only', delete_after=30.0)
 			return
 		async with self.lock:
 			with open(self.filename, 'r') as readfile:
-				lines = readfile.readlines()
+				lines = [line.strip() for line in readfile.readlines()]
 			with open(self.filename, 'a') as appendfile:
 				for word in words:
 					if word not in lines:
 						appendfile.write(word + '\n')
+					else:
+						await ctx.author.send('Duplicate word not added to filter')
+						found_duplicate = True
+		if not found_duplicate:
+			await ctx.message.delete()
+
 
 	@commands.command(
 		name = 'removefilter',
@@ -49,14 +52,14 @@ class WordDetectorCog(commands.Cog):
 	async def removefilter(self, ctx, *words):
 		await ctx.message.delete()
 		if str(ctx.author.top_role) != 'admins':
-			await ctx.author.send('This command is for admins only', delete_after=10.0)
+			await ctx.author.send('This command is for admins only', delete_after=30.0)
 			return
 		async with self.lock:
 			with open(self.filename, 'r') as readfile:
-				lines = readfile.readlines()
+				lines = [line.strip() for line in readfile.readlines()]
 			with open(self.filename, 'w') as writefile:
 				for line in lines:
-					if line.strip() not in words:
+					if line not in words:
 						writefile.write(line + '\n')
 
 def setup(bot):
